init-commit
Implemented the selection of one (by pressing the left mouse button) or several (with the left control held down) Units. Determination of the selected Unit is carried out by hitting the beam into the Unit's collider. All active Units are added to the array. Pressing the right mouse button deselects the Units and removes them from the array.