group-allocation-and-camera-control
Implemented selection of a group of Units with a frame by creating a square and translating its position into world coordinates. When the Units fall into the plane of coordinates of the square, selection and further interaction occurs. Added enumeration of states of interaction with the game. Implemented movement of the camera when clicking or rolling the mouse wheel.