building
Implemented the purchase and construction of buildings. Buildings move on a grid in 1 meter increments. The inspector provides the ability to change the step. If, during the construction of a building, you try to put it on the area of ​​another already built building, then it will light up in red and will not allow you to build it. Comparison of coordinates is carried out using a dictionary.